+++ submodules/TelegramUI/Components/LottieCpp/Sources/Lottie/Private/Model/Layers/SolidLayerModel.hpp
@@ -12,8 +12,8 @@ public:
     explicit SolidLayerModel(lottiejson11::Json::object const &json) noexcept(false) :
     LayerModel(json) {
         colorHex = getString(json, "sc");
-        width = getDouble(json, "sw");
-        height = getDouble(json, "sh");
+        width = (float)getDouble(json, "sw");
+        height = (float)getDouble(json, "sh");
     }
     
     virtual ~SolidLayerModel() = default;
